fix(cluster): overflow-safe entry bounds check + recover in inbound goroutines
V15: decodeEntrySlice computed `off+int(cmdLen)+8 > len(data)` where cmdLen is
uint32; on 32-bit platforms this int expression can overflow and wrap negative,
bypassing the guard and allowing a huge make()/out-of-bounds copy. Compute the
bound in uint64. Add a guard test for the oversized-cmdLen rejection.

V14: inbound cluster goroutines (raft acceptLoop/handleConn, gossip receiveLoop)
lacked panic recovery, so any panic decoding attacker bytes would crash the
whole process. Add defer recover() so a malformed/malicious connection or
datagram tears down only itself; the gossip loop recovers per packet and keeps
running.

+++ internal/cluster/raft/encoding.go
@@ -573,7 +573,10 @@ func decodeEntrySlice(entries *[]entry, data []byte) error {
 		off++
 		cmdLen := binary.BigEndian.Uint32(data[off:])
 		off += 4
-		if off+int(cmdLen)+8 > len(data) {
+		// Bounds-check in uint64: on 32-bit platforms off+int(cmdLen)+8 (cmdLen
+		// is uint32) can overflow int and wrap negative, bypassing the guard and
+		// allowing a huge make()/out-of-bounds copy below (V15).
+		if uint64(off)+uint64(cmdLen)+8 > uint64(len(data)) {
 			return fmt.Errorf("entry slice: command %d overflow", i)
 		}
 		if cmdLen > 0 {
